Removed the react-web folder and moved sub-folders up a level
