+++ test/client/graphql/fragments.test.tsx
@@ -3,10 +3,10 @@ import * as renderer from 'react-test-renderer';
 import gql from 'graphql-tag';
 import ApolloClient from 'apollo-client';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
-import { mockSingleLink } from '../../../../src/test-utils';
-import { ApolloProvider, graphql } from '../../../../src';
+import { mockSingleLink } from '../../../src/test-utils';
+import { ApolloProvider, graphql } from '../../../src';
 
-import stripSymbols from '../../../test-utils/stripSymbols';
+import stripSymbols from '../../test-utils/stripSymbols';
 
 describe('fragments', () => {
   // XXX in a later version, we should support this for composition
