+++ test/client/graphql/mutations/recycled-queries.test.tsx
@@ -3,9 +3,9 @@ import * as renderer from 'react-test-renderer';
 import gql from 'graphql-tag';
 import ApolloClient from 'apollo-client';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
-import { mockSingleLink } from '../../../../../src/test-utils';
-import { ApolloProvider, graphql } from '../../../../../src';
-import stripSymbols from '../../../../test-utils/stripSymbols';
+import { mockSingleLink } from '../../../../src/test-utils';
+import { ApolloProvider, graphql } from '../../../../src';
+import stripSymbols from '../../../test-utils/stripSymbols';
 
 describe('[mutations] update queries', () => {
   // This is a long test that keeps track of a lot of stuff. It is testing
