+++ test/client/graphql/queries/api.test.tsx
@@ -3,16 +3,16 @@ import * as renderer from 'react-test-renderer';
 import gql from 'graphql-tag';
 import ApolloClient from 'apollo-client';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
-import { mockSingleLink } from '../../../../../src/test-utils';
+import { mockSingleLink } from '../../../../src/test-utils';
 import {
   ApolloProvider,
   ChildProps,
   graphql,
   OptionProps,
-} from '../../../../../src';
+} from '../../../../src';
 
-import wrap from '../../../../test-utils/wrap';
-import stripSymbols from '../../../../test-utils/stripSymbols';
+import wrap from '../../../test-utils/wrap';
+import stripSymbols from '../../../test-utils/stripSymbols';
 
 describe('[queries] api', () => {
   // api
