+++ test/client/graphql/statics.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import gql from 'graphql-tag';
-import graphql from '../../../../src/graphql';
+import graphql from '../../../src/graphql';
 
 let sampleOperation = gql`
   {
